+++ config/splat.us.stno1.yaml
@@ -52,6 +52,7 @@ segments:
       - [0x272C, data]
       - [0x34670, .rodata, first_c_file]
       - [0x34670, .rodata, e_red_door]
+      - [0x347BC, .rodata, unk_3CB34]
       - [0x348A8, .rodata, e_collect]
       - [0x34930, .rodata, blit_char]
       - [0x34940, .rodata, e_misc]
@@ -63,6 +64,7 @@ segments:
       - [0x34CBC, .rodata, prim_helpers]
       - [0x34CC4, c, first_c_file]
       - [0x34E6C, c, e_red_door]
+      - [0x3CB34, c, unk_3CB34]
       - [0x419F0, c, st_update]
       - [0x41E24, c, collision]
       - [0x435E0, c, create_entity]
